+++ tabler/tabletypes/csv.py
@@ -35,6 +35,18 @@ class CSV(BaseTableType):
         self.delimiter = delimiter
         super().__init__(extension, verbose=verbose)
 
+    @staticmethod
+    def parse_value(value):
+        """Return None if the value is empty, otherwise return str(value)."""
+        if value == "":
+            return None
+        else:
+            return str(value)
+
+    def parse_row(self, row):
+        """Return a row of parsed values."""
+        return [self.parse_value(value) for value in row]
+
     def open_path(self, path):
         """Return header and rows from file.
 
@@ -42,9 +54,10 @@ class CSV(BaseTableType):
         :type path: str, pathlib.Path or compatible.
         """
         with open(str(path), "r", encoding=self.encoding) as f:
-            csv_file = csv.reader(f, delimiter=self.delimiter)
-            rows = [row for row in csv_file]
-        return rows[0], rows[1:]
+            rows = list(csv.reader(f, delimiter=self.delimiter))
+            header = rows[0]
+            data = [self.parse_row(row) for row in rows[1:]]
+        return header, data
 
     def write(self, table, path):
         """Save data from :class:`tabler.Table` to file.